+++ udata_front/theme/gouvfr/assets/js/pages/Admin/Admin.stories.ts
@@ -106,7 +106,7 @@ export const HomeAdmin: StoryObj<typeof meta> = {
         await delay();
         return HttpResponse.json(requests);
       }),
-      http.get("*/api/1/organizations/6481e3f7959723cf9f8bc98a/datasets/*", async () => {
+      http.get("*/api/1/organizations/646b7187b50b2a93b1ae3d45/datasets/*", async () => {
         await delay();
         return HttpResponse.json({
           "data": [],