+++ udata_front/theme/gouvfr/assets/js/pages/Admin/Datasets/Datasets.vue
@@ -64,9 +64,9 @@
       <div class="fr-col">
         <h2 class="subtitle subtitle--uppercase fr-m-0">{{ t('{n} datasets', totalResult) }}</h2>
       </div>
-      <div v-if="totalResult && oid" class="fr-col-auto fr-grid-row fr-grid-row--middle">
+      <div v-if="oid" class="fr-col-auto fr-grid-row fr-grid-row--middle">
         <div>
-          <a :href="getOrganizationDatasetsCatalogUrl(oid)" class="fr-btn fr-btn--sm fr-icon-download-line fr-btn--icon-left">
+          <a :href="getOrganizationDatasetsCatalogUrl(oid, totalResult)" class="fr-btn fr-btn--sm fr-icon-download-line fr-btn--icon-left">
             {{ t('Download catalog') }}
           </a>
         </div>