+++ udata_front/theme/gouvfr/datagouv-components/src/components/DataserviceCard/DataserviceCard.vue
@@ -1,5 +1,7 @@
 <template>
-  <article class="fr-my-2w fr-p-2w border border-default-grey fr-enlarge-link" :style="props.style">
+  <article class="fr-my-2w fr-p-2w border fr-enlarge-link" :style="props.style" :class="{
+    'border-tabular-api': isTabularApi,
+  }">
     <div class="absolute top-0 fr-grid-row fr-grid-row--middle fr-mt-n3v fr-ml-n1v" v-if="dataservice.is_restricted || dataservice.private || dataservice.archived_at">
       <p class="fr-badge fr-badge--sm fr-badge--mention-grey text-grey-380 fr-mr-1w" v-if="dataservice.is_restricted">
         <span class="fr-icon-lock-line fr-icon--sm" aria-hidden="true"></span>
@@ -16,7 +18,8 @@
     </div>
     <h4 class="fr-text--md fr-mb-0">
       <slot name="dataserviceUrl" :dataservice="dataservice" :dataserviceUrl="dataserviceUrl">
-        <span class="fr-icon-terminal-line fr-icon--sm fr-mr-1v" aria-hidden="true"></span>
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true" class="fr-mr-1v" style="width:1rem" v-if="isTabularApi"><path d="M17.0007 1.20825 18.3195 3.68108 20.7923 4.99992 18.3195 6.31876 17.0007 8.79159 15.6818 6.31876 13.209 4.99992 15.6818 3.68108 17.0007 1.20825ZM10.6673 9.33325 15.6673 11.9999 10.6673 14.6666 8.00065 19.6666 5.33398 14.6666.333984 11.9999 5.33398 9.33325 8.00065 4.33325 10.6673 9.33325ZM11.4173 11.9999 9.18905 10.8115 8.00065 8.58325 6.81224 10.8115 4.58398 11.9999 6.81224 13.1883 8.00065 15.4166 9.18905 13.1883 11.4173 11.9999ZM19.6673 16.3333 18.0007 13.2083 16.334 16.3333 13.209 17.9999 16.334 19.6666 18.0007 22.7916 19.6673 19.6666 22.7923 17.9999 19.6673 16.3333Z"></path></svg>
+        <span class="fr-icon-terminal-line fr-icon--sm fr-mr-1v" aria-hidden="true" v-else></span>
         <AppLink class="inline-flex" :to="dataserviceUrl">
           <TextClamp :auto-resize="true" :text='dataservice.title' :max-lines='1'/>
         </AppLink>
@@ -48,13 +51,14 @@
       v-if="showDescription"
       class="fr-text--sm fr-mt-1w fr-mb-0 overflow-wrap-anywhere"
       :auto-resize="true"
-      :text="dataservice.description"
+      :text="RemoveMarkdown(dataservice.description)"
       :max-lines='2'
     />
   </article>
 </template>
 
 <script setup lang="ts">
+import RemoveMarkdown from "remove-markdown";
 import { useI18n } from "vue-i18n";
 import type { RouteLocationRaw } from "vue-router";
 import TextClamp from 'vue3-text-clamp';
@@ -63,6 +67,8 @@ import type { Dataservice } from "../../types/dataservices";
 import { formatRelativeIfRecentDate } from "../../helpers";
 import OrganizationNameWithCertificate from "../Organization/OrganizationNameWithCertificate.vue";
 import { useOwnerName } from "../../composables";
+import { computed } from "vue";
+import { config } from "../../config";
 
 type Props = {
   dataservice: Dataservice,
@@ -88,4 +94,13 @@ const props = withDefaults(defineProps<Props>(), {
 
 const { t } = useI18n();
 const ownerName = useOwnerName(props.dataservice);
+
+const isTabularApi = computed(() => {
+  return config.tabular_api_dataservice_id === props.dataservice.id
+})
 </script>
+<style scoped>
+.border-tabular-api {
+  border-color: #373C42 !important;
+}
+</style>