+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Current (in progress)
 
+- Add special style to tabular API dataservice card [#548](https://github.com/datagouv/udata-front/pull/548)
 - Export `Placeholder` component, use reuse last_modified [#546](https://github.com/datagouv/udata-front/pull/546)
+- Add missing `dataset` key to `CommunityResource` type [#551](https://github.com/datagouv/udata-front/pull/551)
+- Low enphasis on last element of charts [#562](https://github.com/datagouv/udata-front/pull/562)
+- Remove markdown from dataservice card [#568](https://github.com/datagouv/udata-front/pull/568)
 
 ## 1.4.0 (2024-09-23)